app/requirements.txt
# Install CPU-only PyTorch first to avoid downloading 3.5GB+ of unused NVIDIA CUDA packages
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

transformers==4.46.0
accelerate>=0.26.0
einops==0.7.0
pillow==10.2.0
langchain==0.1.20
langchain-postgres==0.0.6
pgvector==0.2.5
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
streamlit==1.34.0
pypdf==4.2.0